% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A.delicatus_ext}
\alias{A.delicatus_ext}
\title{Afrixalus delicatus genetic groups extent}
\format{
Numeric vector of length length 4
}
\source{
\doi{10.5061/dryad.315km76}
}
\usage{
data(A.delicatus)
}
\description{
Extent for A.delicatus_matrix
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
