% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A.sylvaticus_matrix}
\alias{A.sylvaticus_matrix}
\title{Afrixalus sylvaticus genetic groups}
\format{
A matrix to be converted into a SpatRaster object with a EPSG:4210 projection.
}
\source{
\doi{10.5061/dryad.315km76}
}
\usage{
data(A.sylvaticus)
}
\description{
Raster data representing interpolated genetic group assignments for Afrixalus sylvaticus
based on analyses in Barratt et al. 2018.
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
