% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null_distributions.R
\name{boundary_null_distrib}
\alias{boundary_null_distrib}
\title{Null distribution for overlap statistics}
\usage{
boundary_null_distrib(
  x,
  convert = F,
  cat = F,
  threshold = 0.2,
  n_iterations = 10,
  model = "random",
  projection,
  progress = T
)
}
\arguments{
\item{x}{A RasterLayer object.}

\item{convert}{TRUE if x contains numeric trait data that needs to be converted to boundary intensities. default = FALSE.}

\item{cat}{TRUE if the input RasterLayer contains a categorical variable. default = FALSE.}

\item{threshold}{A value between 0 and 1. The proportion of cells to keep as boundary elements. default = 0.2.}

\item{n_iterations}{An integer indicating the number of iterations for the function.
A value of 100 or 1000 is recommended to produce sufficient resolution for downstream
statistical tests. default = 10.}

\item{model}{Neutral model to use. Options: 'random', 'gaussian_random' (Gaussian random field),
'autoregressive' (simultaneous autoregressive model), and 'random_cluster' (modified random clusters method)}

\item{projection}{Numeric. EPSG code of input raster layer.}

\item{progress}{If progress = TRUE (default) a progress bar will be displayed.}
}
\value{
A list of two probability distribution functions for boundary statistics.
}
\description{
Given a RasterLayer object, this function will simulate a new RasterLayer from a neutral model for a
specified number of iterations. In each iteration, boundary statistics will be computed for the simulated
surface. The function outputs probability functions for two boundary statistics (number of subgraphs
and length of the longest subgraph).
}
\examples{
song_dialect_raster <- raster('song_dialect_boundaries.asc')
null_dialect_boundary <- boundary_null_distrib(song_raster, threshold = 0.2,
                         n_iterations = 1000, model = 'gaussian_random',
                         projection = 4326)

soil_type_raster <- raster('soil_types.asc')
boundary_null_distrib(soil_type_raster, cat = T, n_iterations = 100,
                      model = 'random_cluster', projection = 4210)

}
\references{
Jacquez, G.M., Maruca, I.S. & Fortin, M.-J. (2000) From fields to objects: A review of geographic boundary analysis. Journal of Geographical Systems, 3, 221, 241.
James, P. M. A., Fleming, R.A., & Fortin, M.-J. (2010) Identifying significant scale-specific spatial boundaries using wavelets and null models: Spruce budworm defoliation in Ontario, Canada as a case study. Landscape Ecology, 6, 873-887.
}
\author{
Amy Luo
}
