% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_boundaries.R
\name{categorical_boundary}
\alias{categorical_boundary}
\title{Define the boundary elements of a RasterLayer with categorical data}
\usage{
categorical_boundary(x, projection)
}
\arguments{
\item{x}{A Rasterlayer object.}

\item{projection}{Numeric. The EPSG code of the input layer CRS.}
}
\value{
A RasterLayer object with cell values 1 for boundary elements and 0 for other cells
}
\description{
Creates boundary element cells where patches of two categories meet.
}
\examples{
soil_boundaries <- categorical_boundary(soil_type_raster, projection = 4326)

}
\author{
Amy Luo
}
