% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_boundaries.R
\name{define_boundary}
\alias{define_boundary}
\title{Define the boundary elements of a RasterLayer with numeric data}
\usage{
define_boundary(x, threshold = 0.2, convert = F)
}
\arguments{
\item{x}{A RasterLayer object.}

\item{threshold}{A value between 0 and 1. The proportion of cells to keep as boundary elements. default = 0.2.}

\item{convert}{logical. If TRUE, will convert values of each cell from trait values to boundary intensities. default = FALSE.
value to the estimated rate of change at that cell using a sobel operator. default = FALSE.}
}
\value{
A RasterLayer object with cell values 1 for boundary element and NA for other cell
}
\description{
Defines boundaries in raster objects by keeping a proportion of the cells with the highest
boundary intensity values. If cells have trait values, they can be converted to boundary
values based on approximations of the first partial derivatives of each cell, calculated
using a sobel operator.
}
\examples{
genetic_assignments <- raster('genetic_assignments.asc')
genetic_boundary <- define_boundary(genetic_assignments, threshold = 0.2, convert = T)

genetic_boundary <- raster('genetic_boundary.asc')
genetic_boundary <- define_boundary(genetic_boundary, threshold = 0.1, convert = F)

}
\references{
Fortin, M.J. et al. (2000) Issues related to the detection of boundaries. Landscape Ecology, 15, 453-466.
Jacquez, G.M., Maruca,I S. & Fortin M.-J. (2000) From fields to objects: A review of geographic boundary analysis. Journal of Geographical Systems, 3, 221, 241.
}
\author{
Amy Luo
}
