% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{Ecoregion data for East Africa}
\format{
A RasterLayer object with a EPSG:4210 projection.
}
\source{
\url{https://doi.org/10.1111/mec.14862}. Figure re-digitized into an ESRI shapefile, then rasterized.
}
\usage{
data(ecoregions)
}
\description{
Raster data of ecoregions in East Africa
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
