% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions_crs}
\alias{ecoregions_crs}
\title{Ecoregion data for East Africa projection}
\format{
Projection crs object
}
\source{
\doi{10.5061/dryad.315km76}
}
\usage{
data(ecoregions)
}
\description{
Projection for ecoregions_matrix
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
