% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions_matrix}
\alias{ecoregions_matrix}
\title{Ecoregion data for East Africa}
\format{
A matrix to be converted into a SpatRaster object with a EPSG:4210 projection.
}
\source{
\doi{10.5061/dryad.315km76}
}
\usage{
data(ecoregions)
}
\description{
Raster data of ecoregions in East Africa
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
