% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_random_field_sim.R
\name{gauss_random_field_sim}
\alias{gauss_random_field_sim}
\title{Simulates a Gaussian random field raster with boundary elements.}
\usage{
gauss_random_field_sim(x)
}
\arguments{
\item{x}{A SpatRaster object.}
}
\value{
A SpatRaster object with boundary elements.
}
\description{
Internal function that simulates a raster of the same extent and resolution as
the input raster using a neutral model.
}
\references{
James, P. M. A., Fleming, R.A., & Fortin, M.-J. (2010) Identifying significant scale-specific spatial boundaries using wavelets and null models: Spruce budworm defoliation in Ontario, Canada as a case study. Landscape Ecology, 6, 873-887.
}
\author{
Amy Luo
}
