% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grassland_crs}
\alias{grassland_crs}
\title{Grassland land cover projection}
\format{
Projection crs object
}
\source{
\doi{10.5061/dryad.bk3j9kdhz}
}
\usage{
data(grassland)
}
\description{
Projection for grassland_matrix
}
\references{
Cox et al. (2023) Conservation Genetics
Radoux et al. (2019) Remote Sens 11:354.
}
\keyword{datasets}
