% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grassland_ext}
\alias{grassland_ext}
\title{Grassland land cover extent}
\format{
Numeric vector of length length 4
}
\source{
\doi{10.5061/dryad.bk3j9kdhz}
}
\usage{
data(grassland)
}
\description{
Extent for grassland_matrix
}
\references{
Cox et al. (2023) Conservation Genetics
Radoux et al. (2019) Remote Sens 11:354.
}
\keyword{datasets}
