% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L.concolor}
\alias{L.concolor}
\title{Leptopelis concolor genetic groups}
\format{
A RasterLayer object with a EPSG:4210 projection.
}
\source{
\url{https://doi.org/10.1111/mec.14862}. ADMIXTURE results for samples interpolated to raster with kriging.
}
\usage{
data(L.concolor)
}
\description{
Raster data representing interpolated genetic group assignments for Leptopelis concolor
based on analyses in Barratt et al. 2018.
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
