% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L.concolor_crs}
\alias{L.concolor_crs}
\title{Leptopelis concolor genetic groups projection}
\format{
Projection crs object
}
\source{
\doi{10.5061/dryad.315km76}
}
\usage{
data(L.concolor)
}
\description{
Projection
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
