% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L.flavomaculatus_ext}
\alias{L.flavomaculatus_ext}
\title{Leptopelis flavomaculatus genetic groups extent}
\format{
Numeric vector of length length 4
}
\source{
\doi{10.5061/dryad.315km76}
}
\usage{
data(L.flavomaculatus)
}
\description{
Extent for L.flavomaculatus_ext
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
