% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L.flavomaculatus_matrix}
\alias{L.flavomaculatus_matrix}
\title{Leptopelis flavomaculatus genetic groups}
\format{
A matrix to be converted into a SpatRaster object with a EPSG:4210 projection.
}
\source{
\doi{10.5061/dryad.315km76}
}
\usage{
data(L.flavomaculatus)
}
\description{
Raster data representing interpolated genetic group assignments for Leptopelis flavomaculatus
based on analyses in Barratt et al. 2018.
}
\references{
Barratt et al. (2013) Molecular Ecology 27:4289–4308
}
\keyword{datasets}
