% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa_clusters.R
\name{lisa_clusters}
\alias{lisa_clusters}
\title{Identify LISA clusters}
\usage{
lisa_clusters(x)
}
\arguments{
\item{x}{A SpatRaster object.}
}
\value{
An sf polygons object with LISA clusters
}
\description{
Uses local Moran's I and Monte Carlo simulations to identify LISA clusters
}
\references{
Anselin, L. (1995). Local Indicators of Spatial Association—LISA. Geographical Analysis, 27(2), 93–115.
}
\author{
Amy Luo
}
