% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_statistics.R
\name{max_subgraph}
\alias{max_subgraph}
\title{Length of the longest subgraph}
\usage{
max_subgraph(x, null_distrib, projection)
}
\arguments{
\item{x}{A RasterLayer object with boundary elements.}

\item{null_distrib}{A list of probability functions output from boundary_null_distrib().}

\item{projection}{Numeric. EPSG code of input raster layer.}
}
\value{
The length of the longest subgraph and a p-value.
}
\description{
Statistical test for the length of the longest subgraph, or set of contiguous boundary elements.
}
\examples{
song_raster <- raster('song_dialect_boundaries.asc')
dialect_boundaries <- define_boundary(song_raster, 0.1)
null_dialect_boundary <- boundary_null_distrib(song_raster, threshold = 0.2,
                         n_iterations = 1000, model = 'fractional_brownian',
                         projection = 4326, fract_dim = 1.5)

max_subgraph(dialect_boundaries, dialect_boundary_null, projection = 4326)

}
\references{
Jacquez, G.M., Maruca,I S. & Fortin M.-J. (2000) From fields to objects: A review of geographic boundary analysis. Journal of Geographical Systems, 3, 221, 241.
}
\author{
Amy Luo
}
