% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_random_clust_sim.R
\name{mod_random_clust_sim}
\alias{mod_random_clust_sim}
\title{Simulates raster using modified random clusters with boundary elements.}
\usage{
mod_random_clust_sim(x)
}
\arguments{
\item{x}{A SpatRaster object.}
}
\value{
A SpatRaster object with boundary elements.
}
\description{
Internal function that simulates a  raster of the same extent and resolution as
the input raster using a neutral model.
}
\references{
Saura, S. & Martínez-Millán, J. (2000) Landscape patterns simulation with a modified random clusters method. Landscape Ecology, 15, 661 – 678.
}
\author{
Amy Luo
}
