% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_statistics.R
\name{Odirect}
\alias{Odirect}
\title{Direct overlap between boundary elements}
\usage{
Odirect(x, y, null_distrib)
}
\arguments{
\item{x}{A RasterLayer object with boundary elements.}

\item{y}{A RasterLayer object with boundary elements.}

\item{null_distrib}{A list of functions representing probability distributions. Output from overlap_null_distrib().}
}
\value{
p-value
}
\description{
Statistical test for the number of directly overlapping boundary elements of two traits.
}
\examples{
soil_raster <- raster('soil_types.asc')
genetic_raster <- raster('genetic_assignment_probabilities.asc')

overlap_null_distribs <- overlap_null_distrib(soil_raster, genetic_raster, n_iterations = 1000,
                                              y_convert = T, threshold = 0.1, projection = 4326)
soil_boundaries <- define_boundary(soil_raster)
genetic_boundaries <- define_boundary(genetic_raster)

Odirect(soil_boundary, genetic_boundary, overlap_null_distribs)

}
\references{
Jacquez, G.M., Maruca,I S. & Fortin, M.-J. (2000) From fields to objects: A review of geographic boundary analysis. Journal of Geographical Systems, 3, 221, 241.
Fortin, M.-J., Drapeau, P. & Jacquez, G.M. (1996) Quantification of the Spatial Co-Occurrences of Ecological Boundaries. Oikos, 77, 51-60.
}
\author{
Amy Luo
}
