% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null_distributions.R
\name{overlap_null_distrib}
\alias{overlap_null_distrib}
\title{Null distribution for boundary overlap statistics}
\usage{
overlap_null_distrib(
  x,
  y,
  x_convert = F,
  y_convert = F,
  x_cat = F,
  y_cat = F,
  threshold = 0.2,
  n_iterations = 10,
  x_model = "random",
  y_model = "random",
  projection,
  progress = T
)
}
\arguments{
\item{x}{A RasterLayer object.}

\item{y}{A RasterLayer object.}

\item{x_convert}{TRUE if x contains numeric trait data that needs to be converted to boundary intensities. default = FALSE.}

\item{y_convert}{TRUE if y contains numeric trait data that needs to be converted to boundary intensities. default = FALSE.}

\item{x_cat}{TRUE if x contains a categorical variable. default = FALSE.}

\item{y_cat}{TRUE if y contains a categorical variable. default = FALSE.}

\item{threshold}{A value between 0 and 1. The proportion of cells to keep as
boundary elements. default = 0.2.}

\item{n_iterations}{An integer indicating the number of iterations for the function. A value of 100 or 1000
is recommended to produce sufficient resolution for downstream statistical tests. default = 10.}

\item{x_model}{Neutral model to use for x. Options: 'random', 'gaussian_random' (Gaussian random field),
'autoregressive' (simultaneous autoregressive model), 'fractional_brownian' (fractional Brownian motion),
and 'random_cluster' (modified random clusters method)}

\item{y_model}{Neutral model to use for y.}

\item{projection}{Numeric. EPSG code of input raster layer.}

\item{progress}{If progress = TRUE (default) a progress bar will be displayed.}
}
\value{
A list of probability distribution functions for boundary overlap statistics.
}
\description{
Create a null distribution for boundary overlap statistics, given a pair of variables.

Given two RasterLayer objects with the same range and resolution, this function will
simulate random rasters with boundary elements, then calculate boundary overlap
statistics for n iterations. Statistics for all iterations are returned in a data frame.
}
\examples{
song_dialect_raster <- raster('song_dialect_boundaries.asc')
genetic_raster <- raster('genetic_group_assignments.asc')
null_overlap <- overlap_null_distrib(song_dialect_raster, genetic_raster, y_convert = T,
                                     threshold = 0.1, n_iterations = 1000,
                                     x_model = 'gaussian_random',
                                     y_model = 'gaussian_random',
                                     projection = 4326)

}
\references{
Jacquez, G.M., Maruca, I.S. & Fortin, M.-J. (2000) From fields to objects: A review of geographic boundary analysis. Journal of Geographical Systems, 3, 221, 241.
James, P. M. A., Fleming, R.A., & Fortin, M.-J. (2010) Identifying significant scale-specific spatial boundaries using wavelets and null models: Spruce budworm defoliation in Ontario, Canada as a case study. Landscape Ecology, 6, 873-887.
}
\author{
Amy Luo
}
