% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_raster_sim.R
\name{random_raster_sim}
\alias{random_raster_sim}
\title{Simulates random raster with boundary elements.}
\usage{
random_raster_sim(x)
}
\arguments{
\item{x}{A SpatRaster object.}
}
\value{
A SpatRaster object with boundary elements.
}
\description{
Internal function that simulates a random raster of the same extent and resolution as
the input raster using a neutral model.
}
\references{
James, P. M. A., Fleming, R.A., & Fortin, M.-J. (2010) Identifying significant scale-specific spatial boundaries using wavelets and null models: Spruce budworm defoliation in Ontario, Canada as a case study. Landscape Ecology, 6, 873-887.
}
\author{
Amy Luo
}
