% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rasters.R
\name{simulate_rasters}
\alias{simulate_rasters}
\title{Simulates rasters with boundary elements.}
\usage{
simulate_rasters(x, model, projection)
}
\arguments{
\item{x}{A RasterLayer object.}

\item{model}{Neutral model to use. Options: 'random', 'gaussian_random' (Gaussian random field),
'autoregressive' (simultaneous autoregressive model), and 'random_cluster' (modified random clusters method)}

\item{projection}{Numeric. EPSG code of input raster layer.}
}
\value{
A RasterLayer object with boundary elements.
}
\description{
Internal function that simulates a raster of the same extent and resolution as
the input raster using a neutral model, then defines boundaries in the simulated raster.
}
\details{
This is an internal function.
}
\references{
Saura, S. & Martínez-Millán, J. (2000) Landscape patterns simulation with a modified random clusters method. Landscape Ecology, 15, 661 – 678.
James, P. M. A., Fleming, R.A., & Fortin, M.-J. (2010) Identifying significant scale-specific spatial boundaries using wavelets and null models: Spruce budworm defoliation in Ontario, Canada as a case study. Landscape Ecology, 6, 873-887.
}
\author{
Amy Luo
}
