% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T.cristatus_crs}
\alias{T.cristatus_crs}
\title{Triturus cristatus genetic groups projection}
\format{
ces Barratt et al. (2013) Molecular Ecology 27:4289–4308

Projection crs object
}
\source{
\doi{10.5061/dryad.bk3j9kdhz}
}
\usage{
data(T.cristatus)
}
\description{
Projection for T.cristatus_matrix
}
\references{
Cox et al. (2023) Conservation Genetics
}
\keyword{datasets}
