% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T.cristatus_ext}
\alias{T.cristatus_ext}
\title{Triturus cristatus genetic groups extent}
\format{
Numeric vector of length length 4
}
\source{
\doi{10.5061/dryad.bk3j9kdhz}
}
\usage{
data(T.cristatus)
}
\description{
Extent for T.cristatus_matrix
}
\references{
Cox et al. (2023) Conservation Genetics
}
\keyword{datasets}
