% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T.cristatus_matrix}
\alias{T.cristatus_matrix}
\title{Triturus cristatus genetic groups}
\format{
A matrix to be converted into a SpatRaster object with a EPSG:4326 projection.
}
\source{
\doi{10.5061/dryad.bk3j9kdhz}
}
\usage{
data(T.cristatus)
}
\description{
Raster data representing interpolated genetic group assignments for Triturus cristatus
based on analyses in Cox et al. 2023
}
\references{
Cox et al. (2023) Conservation Genetics
}
\keyword{datasets}
